Awsbinding (#15)
* Added AWS

* Fixed http file

* Added AWS support to Binding Demo

* Updating readme files for AWS binding support

* Added AWS to Secrets demo.

* Working on AWS in PubSub

* changed task type to dapr instead of daprd

* PubSub: AWS Working

* Removed keys from DevOps bicep files.

* Removed keys from Pubsub bicep files.

* Add demos updated to remove keys from bicep.

* Updated the GitHub Workflow.

* Fixed Workflow

* Fixed AWS issues

+++ PubSub/cleanUp.ps1
@@ -9,16 +9,46 @@ param (
     [string]
     $rgName = "dapr_pubsub_demo",
 
+    [Parameter(
+        HelpMessage = "Set to the location of the resources to use."
+    )]
+    [ValidateSet("all", "azure", "aws")]
+    [string]
+    $env = "all",    
+
     [switch]
     $force
 )
 
-# Remove local_secrets.json
-Remove-Item ./components/azure/local_secrets.json -ErrorAction SilentlyContinue
+if ($env -eq 'all' -or $env -eq 'azure') {
+    # Remove local_secrets.json
+    Remove-Item ./components/azure/local_secrets.json -ErrorAction SilentlyContinue
 
-if ($force.IsPresent) {
-    az group delete --resource-group $rgName --no-wait --yes
+    if ($force.IsPresent) {
+        az group delete --resource-group $rgName --no-wait --yes
+    }
+    else {
+        az group delete --resource-group $rgName --no-wait
+    }
 }
-else {
-    az group delete --resource-group $rgName --no-wait
+
+if ($env -eq 'all' -or $env -eq 'aws') {
+    ### AWS
+    # Delete AWS resources
+    if ($(Test-Path ./deploy/aws/terraform.tfvars)) {
+        Push-Location ./deploy/aws
+        $sw = [Diagnostics.Stopwatch]::StartNew()
+        terraform destroy -auto-approve
+        $sw.Stop()
+
+        Write-Verbose "Total elapsed time: $($sw.Elapsed.Minutes):$($sw.Elapsed.Seconds):$($sw.Elapsed.Milliseconds) for deleting a AWS S3"
+        Pop-Location
+    }
+
+    # Remove all terraform files
+    Remove-Item ./deploy/aws/terraform.tfvars -Force -ErrorAction SilentlyContinue
+    Remove-Item ./deploy/aws/terraform.tfstate -Force -ErrorAction SilentlyContinue
+    Remove-Item ./deploy/aws/.terraform -Force -Recurse -ErrorAction SilentlyContinue
+    Remove-Item ./deploy/aws/.terraform.lock.hcl -Force -ErrorAction SilentlyContinue
+    Remove-Item ./deploy/aws/terraform.tfstate.backup -Force -ErrorAction SilentlyContinue
 }
